Fixed many quality errors
