+++ ruleset.ruleset
@@ -8,17 +8,12 @@
     <Rule Id="CA1056" Action="None" />
     <!-- Normalize strings to uppercase -->
     <Rule Id="CA1308" Action="None" />
-
-    <!-- fix -->
     
     <!--CA1032: Add the right constructor to Exception -->
+    <!-- seems fine for custom exceptions on this applications-->
     <Rule Id="CA1032" Action="Info" />
-    
-    <!--CA1034: Do not nest type ReportGeneratorSample. Alternatively, change its accessibility so that it is not externally visible.-->
-    <Rule Id="CA1034" Action="Info" />
 
-    <!--CA1052: Type 'SuccessResult' is a static holder type but is neither static nor NotInheritable-->
-    <Rule Id="CA1052" Action="Info" />
+    <!-- fix -->
 
     <!--CA1305: The behavior of 'int.Parse(string)' could vary-->
     <Rule Id="CA1305" Action="Info" />
@@ -26,23 +21,9 @@
     <!--CA1307: The behavior of 'string.EndsWith(string)' could vary-->
     <Rule Id="CA1307" Action="Info" />
 
-    <!--CA1802: Mark this field as 'const' instead.-->
-    <Rule Id="CA1802" Action="Info" />
-
-    <!--CA1825: Avoid unnecessary zero-length array allocations.-->
-    <Rule Id="CA1825" Action="Info" />
-
     <!--CA2227: Collection properties should be read only -->
     <Rule Id="CA2227" Action="Info" />
 
-    <!--CA2208: Method LoadCoverageFiles passes 'Error' as the paramName argument-->
-    <Rule Id="CA2208" Action="Info" />
-
-    <!--CA2234: Use Uri version of HttpClient methods -->
-    <Rule Id="CA2234" Action="Info" />
-
-    <!--CA5351: Do not use insecure cryptographic algorithm MD5. -->
-    <Rule Id="CA5351" Action="Info" />
   </Rules>
   <Rules AnalyzerId="StyleCop.Analyzers" RuleNamespace="StyleCop.Analyzers">
     <!--SA0001: XML comment analysis is disabled due to project configuration-->
@@ -79,38 +60,10 @@
     <Rule Id="SA1633" Action="None" />
         
 
-    <!-- 
-    <Rule Id="SA1600" Action="None" />
-    <Rule Id="SA1200" Action="None" />
-    
-    <Rule Id="SA1309" Action="None" />
-    
-    <Rule Id="SA1118" Action="Info" />
-    -->
-
-
     <!-- turn on -->
-    
-    <!--SA1003: Operator '(FileInfo[])' should not be followed by whitespace-->
-    <Rule Id="SA1003" Action="Info" />
-
-    <!--SA1122: Use string.Empty for empty strings-->
-    <Rule Id="SA1122" Action="Info" />
     
     <!--SA1202: 'public' members should come before 'private' members-->
     <Rule Id="SA1202" Action="Info" />
-
-    <!--SA1402: File may only contain a single type-->
-    <Rule Id="SA1402" Action="Info" />
-
-    <!--SA1405: Debug.Assert should provide message text-->
-    <Rule Id="SA1405" Action="Info" />
-    
-    <!--SA1502: Element should not be on a single line-->
-    <Rule Id="SA1502" Action="Info" />
-
-    <!--SA1649: File name should match first type name.-->
-    <Rule Id="SA1649" Action="Info" />
 
   </Rules>
   <Rules AnalyzerId="CodeCracker.CSharp" RuleNamespace="CodeCracker.CSharp">
@@ -118,49 +71,13 @@
     <!--CC0001: Use 'var' instead of specifying the type name.-->
     <Rule Id="CC0001" Action="None" />
     
-    <!-- fix -->
-    <!--CC0021: Use 'nameof(fileid)' instead of specifying the program element name.-->
-    <Rule Id="CC0021" Action="Info" />
-
-    <!--CC0022: Process should be disposed.-->
-    <Rule Id="CC0022" Action="Info" />
-
-    <!--CC0031: Verify if delegate 'generateFunc' is null before invoking it. -->
-    <Rule Id="CC0031" Action="Info" />
-
-    <!--CC0091: Make 'UnixifyPath' method static. -->
-    <Rule Id="CC0091" Action="Info" />
   </Rules>
   <Rules AnalyzerId="SonarAnalyzer.CSharp" RuleNamespace="SonarAnalyzer.CSharp">
+  
+    <!-- S2583: Change this condition so that it does not always evaluate to 'false' -->
+    <!-- False positives -->
+    <Rule Id="S2583" Action="Hidden" />
 
-    <!--fix-->
-
-    <!-- S101: Rename class to match camel case naming rules,-->
-    <Rule Id="S101" Action="Info" />
-
-    <!--S125: Remove this commented out code.-->
-    <Rule Id="S125" Action="Info" />
-
-    <!--S1066: Merge this if statement with the enclosing one. -->
-    <Rule Id="S1066" Action="Info" />
-
-    <!--S1075: Refactor your code not to use hardcoded absolute paths or URIs.-->
-    <Rule Id="S1075" Action="Info" />
-
-    <!--S1118: Add a 'protected' constructor or the 'static' keyword to the class declaration.-->
-    <Rule Id="S1118" Action="Info" />
-
-    <!--S1481: Remove this unused 'totalbranches' local variable.-->
-    <Rule Id="S1481" Action="Info" />
-
-    <!--S2583: Change this condition so that it does not always evaluate to 'false'-->
-    <Rule Id="S2583" Action="Info" />
-
-    <!--S3010: Remove this assignment of '_fileSystem' or initialize it statically.-->
-    <Rule Id="S3010" Action="Info" />
-
-    <!--S3928: The parameter name 'Error' is not declared in the argument list. -->
-    <Rule Id="S3928" Action="Info" />
   </Rules>
   <Rules AnalyzerId="Microsoft.CodeQuality.Analyzers" RuleNamespace="Microsoft.CodeQuality.Analyzers">
