+++ src/csmacnz.Coveralls/Data/CoverallData.cs
@@ -28,46 +28,4 @@ namespace csmacnz.Coveralls.Data
         [JsonProperty("git", DefaultValueHandling = DefaultValueHandling.Ignore)]
         public GitData Git { get; set; }
     }
-
-    public sealed class GitData
-    {
-        [JsonProperty("head", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public GitHead Head { get; set; }
-
-        [JsonProperty("branch", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public string Branch { get; set; }
-
-        [JsonProperty("remotes", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public GitRemotes Remotes { get; set; }
-    }
-
-    public sealed class GitHead
-    {
-        [JsonProperty("id", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public string Id { get; set; }
-
-        [JsonProperty("author_name", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public string AuthorName { get; set; }
-
-        [JsonProperty("author_email", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public string AuthorEmail { get; set; }
-
-        [JsonProperty("committer_name", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public string CommitterName { get; set; }
-
-        [JsonProperty("committer_email", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public string ComitterEmail { get; set; }
-
-        [JsonProperty("message", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public string Message { get; set; }
-    }
-
-    public sealed class GitRemotes
-    {
-        [JsonProperty("name", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public string Name { get; set; }
-
-        [JsonProperty("url", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public string Url { get; set; }
-    }
 }
