+++ src/csmacnz.Coveralls/Parsers/ChutzpahJsonFileItem.cs
@@ -1,0 +1,13 @@
+﻿namespace csmacnz.Coveralls.Parsers
+{
+    public class ChutzpahJsonFileItem
+    {
+        public string FilePath { get; set; }
+
+        public int?[] LineExecutionCounts { get; set; }
+
+        public string[] SourceLines { get; set; }
+
+        public double CoveragePercentage { get; set; }
+    }
+}
