+++ src/csmacnz.Coveralls/Ports/ICoverallsService.cs
@@ -1,9 +1,10 @@
+﻿using BCLExtensions;
 using Beefeater;
 
 namespace csmacnz.Coveralls.Ports
 {
     public interface ICoverallsService
     {
-        Result<bool, string> Upload(string fileData);
+        Result<Unit, string> Upload(string fileData);
     }
 }
