--- src/csmacnz.Coveralls/SuccessResult.cs
@@ -1,11 +1,0 @@
-﻿namespace csmacnz.Coveralls
-{
-    public sealed class SuccessResult
-    {
-        private SuccessResult()
-        {
-        }
-
-        public static SuccessResult Value => new SuccessResult();
-    }
-}
